refactor(text-rendering): ♻️ use embeded fonts
Removed the text to path conversion, it was
too complicated to align the generated paths
with the dynamic layouts' texts. So embeded
font files using data URIs is being used,
this is experimental and a "quick-fix", because
data URIs could be not compatible and/or work
in some environments and browsers. A comment and/or
issue with more details about this will be created.